LayoutTests/js/dom/dom-attributes-on-mismatch-type.html

<!DOCTYPE html>
<html>
<head>
<title>Test</title>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
    <div id="testDiv"></div>
    <a id="testAnchor"></a>
    <script>
        description("Make sure DOM attributes handle different types of |this| properly");
        var testDiv = document.getElementById("testDiv");
        var testAnchor = document.getElementById("testAnchor");
        var objectWithDocumentAsPrototype = { __proto__: document };
        shouldThrow("objectWithDocumentAsPrototype.all");
        debug("See what happens if we put the anchor prototype on a div");
        testDiv.__proto__ = testAnchor.__proto__;
        shouldBeTrue("testDiv instanceof HTMLAnchorElement")
        /* This will change to an exception once attributes are on the prototype chain */
        shouldBeUndefined("testDiv.href")
        testDiv.__proto__ = testAnchor;
        shouldBeTrue("testDiv instanceof HTMLAnchorElement")
        shouldThrow("testDiv.href")
    </script>
    <script src="../../resources/js-test-post.js"></script>
</body>
</html>
