LayoutTests/js/dom/dom-attributes-on-mismatch-type-expected.txt

Make sure DOM attributes handle different types of |this| properly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS objectWithDocumentAsPrototype.all threw exception TypeError: Type error.
See what happens if we put the anchor prototype on a div
PASS testDiv instanceof HTMLAnchorElement is true
PASS testDiv.href threw exception TypeError: Type error.
PASS testDiv instanceof HTMLAnchorElement is true
PASS testDiv.href threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
